app/python/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.16.0
google-cloud-logging==3.10.0
